mount, serve, xfer-server: align default port + --nic/--static-ip on mount
Aligns the workstation and UEFI sides on a single default port and
fills the remaining gap in mount's interface-selection knobs.

Port: 9876 instead of 8080.
The previous default collided with everything in the 8080 dev-tool
zone (Jenkins, Tomcat, dev http servers, corp web proxies). 9876 is
IANA-registered to RFC 2974 "sd" (Session Director, defunct in
practice), descending-digit memorable, and far enough from the
crowded range that 'is something else on 9876?' is almost always
no on a corporate laptop. Hardcoded in DEFAULT_SERVER_PORT
(webfs-internal.h), in xfer-server.py's --port default, and in
serve_descs's "port" descriptor. Tests pass an explicit -p 8080 so
they're unaffected. Override via `--port N` on the server and the
URL's `:N` on the mount client.

mount gains --nic / --static-ip.
Until now serve had --nic + --source for interface selection but
mount had no way to pin either side. Adds:

  --nic N           NIC index (default: auto-detect)
  --static-ip A.B.C.D  Static IPv4 to assign instead of DHCP

Deliberately NOT named --source: serve's --source binds the LISTEN
socket (axl_http_server_set listen.ip). Mount has no listen socket;
the analogous knob is the interface's address, hence --static-ip.
Reusing the name would have set up a footgun where the same flag
meant different things on either side of the same project.

DRY: the (uint64_t)-1 → (size_t)-1 nic-index sentinel mapping
that serve_setup did inline now lives once inside network_init.
mount_setup gets to pass the AxlConfig-managed value straight
through with no per-call translation, and serve_setup loses its
copy of the same idiom.

xfer-server.py banner: surfaces the authenticated username on the
Auth line when basic auth is in effect ("Auth: Basic — user
\"mgosha\" (default file ~/.config/axl-webfs/auth)"). Password
never displayed. URL stays bare http://host:port/ so it remains
copy-pasteable into a UEFI Shell mount command without leaking
either half of the credential into terminal scrollback / serial
logs / screenshares.

README updated: port references, banner snippets, and the auth-
enabled banner shape.

Full suite: 100/100 on x64.

+++ src/net/network.h
@@ -21,7 +21,12 @@ typedef struct {
     bool     valid;
 } NetworkInfo;
 
-int  network_init(size_t nic_index, const uint8_t *static_ip, size_t timeout_sec);
+/// Bring up the requested NIC. @p nic_index uses (uint64_t)-1 to
+/// mean "auto-detect" so callers can pass an AxlConfig-managed uint
+/// straight through without a per-caller sentinel translation.
+/// @p static_ip is NULL for DHCP, non-NULL to assign the 4-byte
+/// IPv4 statically. @p timeout_sec bounds DHCP / link-up waits.
+int  network_init(uint64_t nic_index, const uint8_t *static_ip, size_t timeout_sec);
 int  network_get_address(uint8_t ip_out[4]);
 int  network_get_info(NetworkInfo *info);
 void network_cleanup(void);
